requirements.txt
Django==4.1.5
PyMySQL==1.0.2
djangorestframework==3.14.0
python-dotenv==0.21.0